+++ app/controllers/sheets_controller.rb
@@ -61,7 +61,7 @@ class SheetsController < ApplicationController
 
     @raw_data = (params[:format] == 'raw_csv')
 
-    generate_export(sheet_scope, (params[:csv_labeled].to_s == '1'), (params[:csv_raw].to_s == '1'), (params[:pdf].to_s == '1'), (params[:files].to_s == '1'), (params[:data_dictionary].to_s == '1'), (params[:sas].to_s == '1')) if params[:export].to_s == '1'
+    generate_export(sheet_scope, (params[:csv_labeled].to_s == '1'), (params[:csv_raw].to_s == '1'), (params[:pdf].to_s == '1'), (params[:files].to_s == '1'), (params[:data_dictionary].to_s == '1'), (params[:sas].to_s == '1'), (params[:r].to_s == '1')) if params[:export].to_s == '1'
 
     @sheets = sheet_scope.page(params[:page]).per( 40 )
   end
@@ -284,7 +284,7 @@ class SheetsController < ApplicationController
       )
     end
 
-    def generate_export(sheet_scope, csv_labeled, csv_raw, pdf, files, data_dictionary, sas)
+    def generate_export(sheet_scope, csv_labeled, csv_raw, pdf, files, data_dictionary, sas, r)
       export = current_user.exports.create(
                   name: "#{@project.name.gsub(/[^a-zA-Z0-9_]/, '_')}_#{Date.today.strftime("%Y%m%d")}",
                   project_id: @project.id,
@@ -294,6 +294,7 @@ class SheetsController < ApplicationController
                   include_files: files,
                   include_data_dictionary: data_dictionary,
                   include_sas: sas,
+                  include_r: r,
                   sheet_ids_count: sheet_scope.count )
 
       rake_task = "#{RAKE_PATH} sheet_export EXPORT_ID=#{export.id} SHEET_IDS='#{sheet_scope.pluck(:id).join(',')}' &"