+++ db/migrate/20140904192001_add_include_r_to_exports.rb
@@ -1,0 +1,5 @@
+class AddIncludeRToExports < ActiveRecord::Migration
+  def change
+    add_column :exports, :include_r, :boolean, null: false, default: false
+  end
+end