+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Option values no longer need to be entered and will fall back to the default placeholder numbers if left blank
 - **Export Changes**
   - The SAS export now applies labels and domains to checkbox subvariables
+  - An R export option has been added to the statistical export section (thanks to @mcailler for the template R script)
 - **Adminstrative Changes**
   - System admins can now see if a user has been invited to a site to aid in the user activation process
 - **Gem Changes**