+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.26.0
+## 0.26.0 (September 6, 2014)
 
 ### Enhancements
 - **Auditing Changes**